Initial commit
With this commit I created a new extension which brings Doctrine DBAL to
TYPO3. Here I just copied the relevant classes from core into this
extension for later changes. In this state it is already possible to
select this extension in install process but it behaves exactly like the
core for now. Lets change it!