[WIP] Refactoring the old and the new API
The class with the old methods extends the class with the new methods.
When its useful it uses the new methods, otherwise it uses his old
code.

This brings also an interface with.

The legacy code is marked as deprecated and the whole class could just removed
when the time it has come.