[Postgres] Provide the table name to get the last inserted ID
While MySQL is very simple to get the last inserted ID of a dataset,
Postgres need a sequence to do so. A sequence name has the pattern
"tablename_primaryindexfield_seq" -> "pages_uid_seq". To make it easy
for the developer we just need to pass the name of the table and the method
will construct the name of the sequence for us.

But this includes to have in every table a field with the name "uid"
which is the primary key.

Would be nice to find a better solution for this, so that we not have to pass
the table name.